clippy

+++ crates/wgsl_analyzer/src/from_proto.rs
@@ -25,13 +25,13 @@ pub(crate) fn vfs_path(url: &lsp_types::Url) -> Result<vfs::VfsPath> {
 pub(crate) fn offset(line_index: &LineIndex, position: lsp_types::Position) -> Result<TextSize> {
     let line_col = match line_index.encoding {
         OffsetEncoding::Utf8 => LineCol {
-            line: position.line as u32,
-            col: position.character as u32,
+            line: position.line,
+            col: position.character,
         },
         OffsetEncoding::Utf16 => {
             let line_col = LineColUtf16 {
-                line: position.line as u32,
-                col: position.character as u32,
+                line: position.line,
+                col: position.character,
             };
             line_index.index.to_utf8(line_col)
         }
